Split core si-id macro into multiple macros
This commit splits the core si-id macro into multiple macros and is a
follow-up to #5093.

The macros are based on the existence of PostgreSQL conversion
implementations and the "nil" ID. The macros using "nil" should not be
used by any new IDs and exist for backwards compatibility.

Additional changes:

- All macros have moved into their own module
- All conversions between IDs have been moved into their own module
- Our standard set of lints for new crates have been added to "si-id"
  (along with resolutions to failed lints)

Potential future work:

The below list has been carried over and modified from from #5093:

- Migrate all IDs in "si-layer-cache" (currently not reliant on a macro)
- Migrate all IDs in "module-index-server" (currently not reliant on a
  macro)
- Clean-up ID call sites (e.g. reduce "pub use" statements, but we may
  not actually want this)
- Remove all usages of the third party ulid crate and go through "si-id"
  (we may not actually want this)
